requirements.txt
dateutils==0.6.6
evernote==1.25.3
httplib2==0.11.3
lxml==4.2.1
Mako==1.0.7
MarkupSafe==1.0
oauth2==1.9.0.post1
python-dateutil==2.7.2
pytz==2018.4
six==1.11.0
